Adds a wrapper function for IORegistryEntry property getting boilerplate
Reading a property of a specific type from an IORegistryEntry object requires a bunch of code for checking against nullptr, casting to the correct object type, and finally extracting the data.

This change adds a djt_get_registry_entry_property_number() function for doing this song and dance in one place anytime you need to read a number-typed property.