New gizmo: Fixup for dexts with new SDK crashing on older macOS versions
This change adds a function to call in user client subclasses for preventing dexts built with DriverKit SDK 21 or newer from crashing on macOS 11 or older.

The root cause for the crash is that the IIG definition for IOUserClient places the ExternalMethod function on a named dispatch queue. (kIOUserClientQueueNameExternalMethod) The DriverKit 21 runtime apparently sets this queue to the default queue automatically, but the runtimes in macOS 11 and older are unaware of this requirement. On those versions the queue must therefore be explicitly set or queue name lookup fails with a failed assertion.